refactor: LinkedListQueue타입을 LinkedList타입으로 축소 * LinkedList 타입을 class에서 구조체로 변경 * append, checkFirst, remove, removeAll 로 메서드 이름 변경 * count를 증감시키는 메서드 분리 구현
